Re-introduce support for mcrypt
It's not quite clear how mcrypt ever worked, but the symptoms suggest
that previous versions of this build pack were built on another Ubuntu
machine (as evidenced by use of apt-get install) and that PHP of that
era may have statically linked mcrypt.

This effectively vendored in mcrypt, even though such an intention is
not evidenced anywhere in the buildpack's history.

Subsequent builds that have been tweaked to work entirely in-Dyno have
thus lost this opportunity to statically link mcrypt, and furthermore
there are accounts on the web that mcrypt is dynamically linked
nowadays[0].

So, vendor mcrypt, ./configure-ing it as described by the manual
comments.

Furthermore, since this changes the PHP binaries even though it's the
same PHP version (the ./configure flags are different), introduce the
notion of "heroku_rev" that is used to signify such
sub-upstream-release versioning.  For simplicity, it is applied
uniformly to all emitted binary artifacts of the build, even though
some shouldn't change (in this case, Apache has received no changes
but has had its heroku_rev incremented anyway).

[0]: http://www.php.net/manual/en/mcrypt.installation.php