Uploading binaries immediately after compiling ----------------------------------------------
WARNING: One ought not log out of the 'heroku bash' shell after compiling-- doing so would destroy the files that just took 20 minutes to compile!  This is both the curse and (in performance) the magic of the ephemeral filesystem.

Instead, one can run the following as executable text, in order to whisk those files out to Amazon S3 immediately.