Move sleep outside of httpd/nginx subshell
This used to be in the httpd/nginx subshell because if it was between the FPM/HHVM subshell and the httpd/nginx one, FPM/HHVM would "hang" (inside the first subshell!) during the sleep, not continue initializing (which is the whole point), and "resume" together with httpd/nginx (which runs in its own subshell); really weird. This rendered the sleep useless - the point is to give FPM a moment to finish initializing before sending traffic to it, or else Nginx will mark it as unhealthy under load with the first few requests hitting errors.
After the removal of the redirect outside the subshell (in 7c30f7c), this now suddenly works again. A bit cleaner this way, so let's do it. Completely unclear why the redirect is related and how this can even happen, though.