work around internal Composer package ordering
Composer uses only name and version for resolution of a package. For extensions, we have multiple packages with identical name and version, but different entries for 'php' in 'require'. Composer will pick the first match for resolution and continue, which means that if a user requires '^5.5.10' for 'php' and '*' for 'ext-memcached', then PHP 5.5.something will be picked if the 'ext-memcached' package that has '5.5.*' as its PHP requirement occurs before the 'ext-memcached' package with the same version that has '5.6.*' as its PHP requirement.

We simply sort the repo by 'php' or 'hhvm' requirement in descending order to work around this.