fix automatic ext installs failing on PHP 7
On build, the blackfire or newrelic add-ons get detected, and then a composer require is attempted for the platform packages.
With many deps, our generated composer.json triggers PCRE backtracking (all versions) or JIT stack size limit (in PHP 7) errors because Composer does mad regexes that fail with a lot of JSON data before the "require" block.
This change moves the "require" section, which composer modifies on "composer require", to before the big list of "repositories".