Don't suppress FPM errors
When using php in a frontend application it's relatively fine, but in an API context this is a much bigger issue.
Essentially what happens is nginx sees any output from php-fpm as valid, even though it's not, and will return 200 ok. 

for reference: https://ma.ttias.be/nginx-sets-http-200-ok-php-fpm-parse-errors/

By default, it should fail, not provide a false sense of security.