Default to listen.mode=0666 for PHP-FPM socket
This is to allow running in both Heroku Dynos and containers.

As both dynos and containers only run a single site/pool, a mode of 0666 is okay.

A solution via listen.owner is not possible, because the dyno user can't chown the socket to e.g. nobody or daemon user ownership.