Restructure Nginx configs and loading
A main.conf is now bundled with the buildpack, as well as updated fastcgi_params and mime.types.

This main config is used for the server via -c (instead of the system default), and the previous "heroku.conf.php" config still gets included using a dynamic directive.

This retains backwards compatibility for existing custom configs, but allows Nginx versions 1.9.3 and newer, which do not allow more than one "http" block, to work.

Because changing the base config with "nginx -c" also resets the directory relative to which "include" lookups occur, we must bundle "fastcgi_params" and "mime.types", as they may be included in existing "heroku-php-nginx -c" or "heroku-php-nginx -C" situations.

Fixes #198