Stop using php -n to prevent APM extension startup
We do not want New Relic and others to start and instrument during simple checks and processing and Composer invocations on boot.

The previous approach with php -n prevents all PHP INIs from getting loaded, which means setups where even basics like PCRE or JSON are shared extensions and are loaded via a php.ini config directory would not work.

This change instead appends a directory to PHP_INI_SCAN_DIR, and that directory contains an INI file with directives that prevents the startup of common APM extensions.

Fixes #345, #348, #349