Support `python` not being symlinked to `python2`
Historically the Ubuntu `python2.7` package has symlinked `python` to
`python2`. However starting in Ubuntu 20.04, this symlink is no longer
created, and one must explicitly call `python2` (or `python3`) instead.

Unfortunately we cannot just `s/python/python2/` throughout, since this
buildpack aims to also support OS X for local development workflows, and
on OS X system Python (which is Python 2) is only available as `python`
and not also as `python2`. Whilst homebrew Python does have `python{2,3}`
binaries, we don't want to require use of it.

As such, for the scripts that may be run locally (`bin/heroku_*`), they
must now check for the presence of either `python2` or `python`.

This change is a no-op for current stacks, but is landing now to reduce
the size of the Heroku-20 support PR.

In the future we'll want to convert these usages to Python 3, however
we are explicitly not doing so for now, since:
* this work is holding up the Heroku-20 release, and a no-op change is
  safer to land quickly than changing version
* Python 3 is not available out of the box on OS X, so we'd still have to
  use Python 2 there, which would mean needing to test against multiple
  Python versions in CI to ensure continued Python 2 compatibility. This
  is doable, but should not be on the critical path for releasing
  Heroku-20.

Refs W-7491249.