Inject `watiforit.sh` using `before_deploy` append
This monkey patching of hatchet is causing problems because the behavior of `setup!` was changed to not run any commands due to a bug.

This broke the expectations that were being used to patch `setup!`. Instead, there's a supported interface to use instead: `before_deploy(:append)` that was just added and released in Hatchet version 7.3.0.