Skip HEROKU_PHP_PLATFORM_REPOSITORIES for cache behavior test
When changing the stack, we'd also have to change HEROKU_PHP_PLATFORM_REPOSITORIES to point to the correct repo for that stack.

This env var differs between main/tags and other branches - main/tags use stable binaries, all branches use staging binaries.

We don't care about the binaries for this test, since we're testing buildpack behavior, so it's fine to just skip that stack/platrepo helper altogether and have the buildpack use its default repo for the run.