Use our own Heroku CLI install, disable autoupdate
Travis has the CLI pre-installed, but it's not always quite up-to-date, leading to hundreds of warning messages from the CLI that an update is available.

Hatchet itself won't install the CLI if it's already there, so we're doing it by hand here, and using the non-Ubuntu install, because that overwrites the Travis provided version in /usr/local/bin (the Ubuntu installer version wouldn't take precedence in /usr/bin), and because it's much faster.

This is also much faster than doing a "heroku update", which downloads half of the Internet in the form of NPM metadata.

We're also setting an env var to disable auto updates for the CLI, since we just did a fresh install of the latest release.