Fix ext-newrelic require printing logs on build
Our standard log level for NR is warning, but this is only applied at runtime. During builds, this usually doesn't matter, as the extension is enabled magically after the build, but if there is an explicit ext-newrelic require, then the build output will be littered with NR agent logs saying it's disabled, which is confusing.

Tests are adjusted accordingly; they do rely on "info" level logging so we can assert that the agent is in fact not starting up etc, so we're adding an explicit case where we use the default of "warning" to ensure the log lines are not there in that case.

GUS-W-8215073