Ignore $WEB_CONCURRENCY if $WEB_CONCURRENCY_SET_BY is set
If there is a `$WEB_CONCURRENCY_SET_BY` value, it means another buildpack ran after us, and we are seeing its calculated value for `$WEB_CONCURRENCY` (as that later buildpack overwrote our empty `.profile.d/WEB_CONCURRENCY.sh`).

In that case, we print a notice, and ignore the variable.

Also, we set `$WEB_CONCURRENCY_SET_BY=heroku/php` in into the web environment if we calculate the concurrency, just for consistency with other buildpacks.

GUS-W-20838926