PHP 8.1 compat for extension installs
For extensions, PHP is a build-time dependency, but both PHP and the extension end up under the same install prefix.

In order to not have PHP itself in the extension archive, the common pattern is to 'make', then to 'rm -rf'  the build dependencies, then to 'make install' the built extension - the build result is only moved into the prefix path during the install.

In PHP 8.1, however, many core PHP headers are listed as dependencies for the Linker in the Makefile generated by 'phpize', so we need at least those for an installation.

As our install also installs headers into the same 'include/'  structure, however, we need to change our approach for a clean, self-contained install for archiving.

PHP's generated Makefile allows for an INSTALL_ROOT environment variable to override the prefix during installation, so we can use that instead.