Forward compatibility for Composer installs
Right now, we read the Composer API version from the user's lock file, and use that to determine the version of Composer for the userland dependency install step.

The problem is that every Composer Minor (2.x.0) release bumps this version now, so e.g. recently, when 2.2.0 came out, the lock file said 2.2.0, we used ^2.2.0 to pick a Composer version, and that then failed for users, as we had't made Composer 2.2 available yet, see e.g. https://github.com/heroku/heroku-buildpack-php/issues/526

We now instead only pick the major version component from that version selector, as Composer is more or less guaranteed to always be forward compatible - Composer 2.0.0 can still install a lock file generated by 2.2.3, and the maintainers confirmed to me that this will remain so: https://twitter.com/seldaek/status/1474477031329050625

GUS-W-10372327