Shared PHP extensions are now separate packages
Shared extensions bundled with PHP have special handling in the installer. The reason is that if a composer.json requires "ext-bcmath" (bundled with PHP), Composer will not fire a separate install event, as the previous installation of package "php" already satisfies the requirement. So the installer plugin maintains a list of all extension requirements, and, when installing a PHP package, manually enables these extensions by writing an INI config for PHP to load the shared module, just the same way it would for a separate "external" extension package.

Aside from adding complexity to the code (which isn't removed yet, as it would be a BC break for people using third-party repositories, and because only our maintained PHP versions get rebuilds with this change), this leads to an inconsistency with how so-called "polyfills" are treated.

Polyfills are userland packages that claim to "provide" a native extension, by implementing the functionality in userland. Popular examples are e.g. https://packagist.org/packages/symfony/polyfill-mbstring or https://github.com/phpseclib/mcrypt_compat for extensions that are (or at some point were) bundled with PHP; https://packagist.org/packages/symfony/polyfill-uuid or https://packagist.org/packages/alcaeus/mongo-php-adapter are polyfills for third-party extensions.

When these are present, the native extension, if it were a separate package, would not be selected for installation by Composer's dependency solver. However, our installer plugin can't know that - it only knows that somewhere in composer.json, "ext-mbstring" is required, and that package "php" says it "replace"s that package, so it gets enabled. It can't know that "symfony/polyfill" claims to "provide" a suitable version of "ext-mbstring".

That behavior does not lead to problems, as the polyfills always contain logic to fall back to the native extension they're "emulating" in userspace if it's present, but it is inconsistent with how the installer behaves for extensions that are "real" packages, e.g. "ext-uuid", which will not be installed or enabled when https://packagist.org/packages/symfony/polyfill-uuid is used.

This change causes the repository generator to expand bundled shared PHP extensions into their own packages, with a "dist" "type" that will cause the installer to not download anything. The packages will be treated like any other extension by Composer for dependency resolution, and will go through a normal installer event inside Composer, meaning the same logic for enabling extensions as for third-party extensions is used.

In a future version of the installer, the legacy logic for handling shared bundled PHP extensions can then be removed.

GUS-W-9889113