Extensions declare "native" replaces of themselves
All extensions, whether bundled as shared with PHP, or third-party, are now standalone packages, and the Composer solver will not select them if a userland polyfill package "provide"s them.

This does not lead to the best performance, and cannot usually be avoided by users of e.g. Symfony or Laravel, as these frameworks use the https://github.com/symfony/polyfill packages as internal dependencies.

In preparation of a future change that attempts forceful enabling of extensions provided by polyfills, this change causes all extensions to also declare a "heroku-sys/ext-EXTNAME.native" variant of themselves (as "replace"d), and PHP to declare all statically compiled-in extensions the same way.

With an upcoming change, this will then allow us to essentially try to "composer require heroku-sys/ext-EXTNAME.native" for every EXTNAME that is "provide"d by a polyfill, with the guarantee that the rest of the dependency graph will not be modified in order to satisfy the dependency.

Including this change early to avoid re-building the entire repository, as this is part of the 1.6 installer spec already.

GUS-W-9889127