Install extensions provided by userland packages
Some userland packages declare extensions as "provide"d in their manifest. If such a userland package is in the dependency graph, the actual native extension will not be selected during solving by Composer.

Sometimes, that's fine (e.g. for extensions that are no longer available for recent PHP versions); sometimes, that means slower performance or, worse, malfunctions (if a "polyfill" only implements a subset of the extension's features).

Since all of our extensions (and PHP, for its statically compiled in extensions) now declare ".native" variants as "replace"d in their manifest, we can attempt to install those variants after the initial platform install.

That will not always work (e.g. if the package is not available at all, or if Composer would have to downgrade or remove an installed package to satisfy the requirement), but that's totally fine - in that case, the polyfill works exactly as designed.

GUS-W-9889127