use Composer 2.2 for older lockfiles
Composer records the plugin API version (2.2.0, 2.3.0 etc) of the Composer release (e.g. 2.3.2) that generated the lock file inside composer.lock.

Composer 2.3 changes a lot of internal APIs, so projects with Composer plugins may not be compatible. It also updates the Symfony components it uses from 2.8 or 5.4, which can cause problems for old PHP-based install hooks.

Since such plugins or hooks often do not specify a compatible version range, or followed semver with e.g. ^2.0.0, they would still install with 2.3.0, and then fail.

As Composer 2.2 is an LTS series, we will simply use Composer 2.2 if the lock file was generated using Composer 2.0, 2.1 or 2.2.

GUS-W-10940526