Do not output Composer 'No composer.lock file present...' message on platform install failure
Composer prints this as part of its regular output, but there have been tickets where both customers and support agents were confusied by this output.

It refers in this case to the platform installer lock file, not the userland file.

This change drops the entire line from the composer install output.

GUS-W-11272794