Fail if platform packages are part of app source
Sometimes, users download a slug from Heroku, and put that slug into a Git repo.

As a result, the source will contain hundreds of megabytes of binaries in .heroku/php/, and platform installation will fail because Composer things nothing needs to be installed (which is correct, but the post-install hooks that set up $PATH etc will not trigger).

Better UX to fail the build in this case.

Also gives a human-readable error now in case the buildpack is set twice on the app for some reason.

GUS-W-12004810