Support CNB package installs in ComposerInstallerPlugin
For CNBs, packages can now declare an `install` script in `extra.heroku` (similar to `export` and `profile` for Classic) in the package metadata that is executed by the installer plugin after the package was installed.

In there, operations such as e.g. "relocating" PHP or Apache to the different install prefix can be done using scripting.

For this purpose, the script is called the root of the destination install path as an argument (e.g. `/layers/heroku_php/platform`, with a trailing slash).

The script is expected to print a JSON structure to stdout that will be read and applied by the CNB to the LayerEnvData for the layer; this can be used to set the PATHs that each package needs.

For example, a `bin/install.mypackage.sh` could look roughly as follows for a package like PHP:

    #!/usr/bin/env bash
    set -eu

    package_root=$(cd "$(dirname "$0")"; cd ..; pwd)
    dest_root=$1

    # install prefix was /app/.heroku/php
    sed -i "s#/app/.heroku/php#${dest_root}#g" "${package_root}/etc/php.ini"

    cat <<-EOF
    [
      {
        "scope": "all",
        "modification_behavior": "prepend",
        "name": "PATH",
        "value": "${dest_root}/sbin"
      },
      {
        "scope": "all",
        "modification_behavior": "prepend",
        "name": "PATH",
        "value": "${dest_root}/bin"
      },
      {
        "scope": "all",
        "modification_behavior": "delimiter",
        "name": "PATH",
        "value": ":"
      },
      {
        "scope": "all",
        "modification_behavior": "override",
        "name": "PHP_INI_SCAN_DIR",
        "value": "${php_ini_scan_dir}"
      },
      {
        "scope": "all",
        "modification_behavior": "override",
        "name": "PHPRC",
        "value": "${php_ini_path}"
      }
    ]
    EOF

GUS-W-21580958