Move handling of ".native" extension variants to mkrepo.sh, fix some edge cases, add tests
With this change, package formulae no longer need to care about the `.native` package variant trick we use to install native extensions that are also declared as provided by polyfill packages - all metadata is produced during repository construction.

This means that extension packages will now also always automatically `require` the `.native` variant of another extension they depend upon - this was previously not the case, and e.g. a userland polyfill of `ext-psr` would have not gotten installed before the `ext-phalcon` that depends on it.

This is not a problem in that particular case (or the other one we have, where `ext-pq` requires `ext-raphf`), since the native package would get installed in the "try `.native` installs" step after the main platform install, but there are cases where extension load order matters (for symbol definitions), and, more importantly, we absolutely must ensure that any native dependencies are actually there - if `ext-pq` needs `ext-raphf`, but `ext-raphf`, for whatever reason, can't be installed, then `ext-pq` would not function.

GUS-W-13772974