Fix waitforit.sh hanging on Focal
New Focal based instances spawn bash in a way that does not allow the installation of SIGPIPE handlers.

But waitforit.sh needs SIGPIPE to detect when a test finishes. The workaround is to now manually issue a different signal when the write to a popeline fails.

This would cause certain tests to hang forever.

Put some comments in to better explain how the whole thing works.

GUS-W-14213860