Fix matching of extension requirements against PHP series
There are two kinds of extensions: those bundled with PHP, and those that are separate packages.

The former come in two flavors: bundled and compiled in, or bundled and built as shared.

For compiled in extensions, entries are generated in the `replace` section of the PHP release they belong to.

For shared extensions, standalone "virtual" packages are generated during repository generation. They will have a `require` entry for the exact PHP version they are bundled with, since they aren't really packages that get downloaded and installed, but instead only activated in their "parent" PHP package upon installation.

As a result, the right package choices are made by the Composer solver during platform package installation.

For third-party extensions, the `require` entry for `heroku-sys/php` does not specify an exact version, but a selector like "`8.2.*`", since a build of e.g. `ext-amqp` version 2.1.1 for PHP series 8.3 is compatible with all 8.3.x builds (due to PHP's extension API and ABI compatibility).

This has evolved over time, and the Dev Center generator table only happened to be correct at this point due to the way the extension's requirements for `php` were matched against the known list of PHP series - an "`8.2.*`" from a third-party extension would be satisfied by a series entry (those look like "`8.2`"), but an `8.2.15` would not match. However, "`8.2.0`" would match "`8.2`", and then all hell breaks loose in the internal database used during generation, and the output results are all over the place (built-in extensions showing up in the third-party list, third-party extensions listed as unavailable even if they're there, etc).

This change corrects the loading of and metadata generation for bundled extensions.

GUS-W-14625659