Generate only whitelisted PHP series info
Instead of producing runtime rows, and extension columns, for PHP series that are not whitelisted, we throw a warning.

Now also prints a notice if a whitelisted runtime series isn't in the input (which can happen when e.g. only generating for a single stack, or when a stack goes EOL).

GUS-W-14625693