Simplify autotune.php output/help a little
Do not output CPU core info unless in verbose mode, but do print a notice when the number of processes gets limited to below RAM_AVAIL/memory_limit.

This reduces the potential for confusion (think number of cores on shared dynos), and keeps the printed info consistent with prior versions of the buildpack.