Simplify waitforit.sh
The purpose of this 'program' is to wait for output from a launched program, but with a timeout to prevent hanging forever.

Its key feature is the ability to detect when it is part of a pipeline, and to then write something to that pipeline to signal that the expected output is there - this allows constructs like 'waitforit.sh 10 someprogram | { read; do_something_with_someprogram; }'.

The previous solution for the internal pipeline signaling is, however, subject to certain race conditions, and then programs might not be terminated correctly, or the pipe signal loop may hang forever.

It's simpler overall to just launch two 'timeout's, for the program and for the output grepping, and wait for either or both to terminate.