Improve handling of concurrent or unknown signals
When we are, for whatever reason, sent the same signal multiple times, the very first operation inside the trap ('trap - ...') currently re-sets the signal handler (to SIG_DFL). Since Bash 4.3 RC2, delivery of signals is no longer blocked during the execution of traps for those same signals (i.e., the same trap handler that is executing fires again during its own execution). Since we've re-set to SIG_DFL at the beginning of the trap, if another TERM signal arrives while the trap is running, the default behavior of immediate termination kicks in, which also kills the currently running trap and its cleanup procedure.

Instead, we now perform a 'trap "" ...' to instead ignore this signal immediately (SIG_IGN), perform the cleanup, and then re-set the trap to the default handler before triggering it by sending the signal to ourselves - this is the correct behavior for a program that terminates in response to a signal in order for other programs to be able to correctly determine the cause of a termination (so e.g. when receiving SIGTERM, a program must, as its final action, kill itself using that signal, and not exit with e.g. 0).

This means that the subshells will now exit with a non-0 exit status; this triggers an error condition in the main shell, which fires off the signal handler for the signal received - so a subshell that terminates with 'kill -USR1 $$' in response to a USR1 signal will have an exit status of 128+USR1, and trigger a USR1 signal in the main shell. That's why we ignore that USR1 signal before issuing the cleanup, since it would otherwise trigger the (now reset) EXIT handler, and terminate the outer shell.

This behavior is quite useful for any unhandled signals now - if the main program or a subshell receive e.g. a USR2, which by default causes termination, the main EXIT trap fires and tears stuff down.

There is a possible race condition in Bash's delivery of signals to trap handlers that will occasionally cause the internal trap list to become corrupted if signals arrive in rapid succession between the execution of a trap and the changing of a signal handler (to SIG_IGN in our case). This is often, but not always, accompanied by a message like "warning: run_pending_traps: bad value in trap_list[15]", and it will happen if signals arrive in rapid succession (see run_pending_traps() in Bash's trap.c).

If this occurs, the re-setting of the trap to SIG_DFL, and the subsequent termination due to the 'kill -$SIGNAL $$', will not execute (and the bash process will hang indefinitely) until the trap exits for another reason. We "trigger" this exiting with an explicit 'exit 0' call in the outer shell's traps (in this case, the exit 0 will not actually execute, and the termination status of the outer process will be the correct e.g. "143" status code representing termination in response to SIGTERM). In the subshells, this is not necessary, because they get cleaned up by the outer bash process later.

GUS-W-15071619