Also wait in subshell exit traps
This is useful when unhandled signals arrive - the write to fd3 will unblock the read in the main script, and start the cleanup process, which will then collect us, since we're still wait-ing on our subprocess.

Without this, issuing e.g. a SIGUSR2 to a subshell would cause termination of everything else, but not the children of this subshell.

GUS-W-15071619