No need to re-set EXIT traps in subshells
Worst case, this would print an unexpected exit message in the parent, which is kind of useful: it tells folks that there is signals sent not only to the leader.

However, it seems that bash does not execute an EXIT trap anyway in this, due to the 'kill -TERM $$' at the end of the trap (which is correct behavior).

Either way, the flow now for a HEROKU_PHP_GRACEFUL_SIGTERM=0 execution and a "carpet bombed" SIGTERM delivery to all processes is now for the main and the subshell TERM traps to fire in random-ish order; eventually, one of the subshells will terminate with its 'kill -TERM $$', which, due to set -e in the outer shell, will actually fire the TERM trap again, which is already in progress and gets ignored (or, if it is not already ignored, because somebody signaled TERM only to a subshell, this will trigger the cleanup procedure, and tear everything else down).

GUS-W-15071619