Drop 'publish' input from platform-build.yml
We always want this to happen.

This option used to make sense when generating a repo was a multi-minute affair, but now it really is redundant; we do not want deploys of packages to S3 that aren't written to packages.json but are then picked up by random other runs of mkrepo.sh.