Work around Bash buffering bug causing test failures
The boot tests run several iterations inside a single 'heroku run', as rthis drastically reduces the overall time taken for tests.

Some Bash 5.0 bug (it seems to happen only on heroku-20) causes output from a command (that is still running through a tee process substitution) to arrive a bit after the next echo statement, and as a result, sometimes, the last line of a delimited chunk is not the exit status, but output from the program, causing the test to fail.

So we just write outputs to files, and cat them together at the end.