Allow latest Composer if no lock file present
We're bumping the plugin-api-version for the no lock file present case to allow Composer 2.8 which works with PHP 8.4 (to allow clean boots, and builds without deprecation errors).

This is a very inconsequential change for users that have no lock file, as no lock file means no package installs, plugins, etc.