PHP/8.4.1
There is no more E_STRICT for error_reporting, and session.sid_length is gone.

The imap extension has moved to PECL.

No newrelic extension builds yet with support for 8.4. The phalcon extension is incompatible with 8.4 (they always get new releases some time later).

The heroku-20 stack is not getting this version because the stack is already deprecated.

GUS-W-17226361