Minor cleanup for platform repo priorities tests and docs
Fix the repos (and the build docs examples) to have objects with version arrays for each package, as Composer prefers it (and how all the tooling has been generating it for a while). No big thing, as the 'old way' still works (but throws a warning, which doesn't affect the tests though).

Also restructure the corresponding test to have a sub-group that describes which test case ran.

GUS-W-17622481