Human-readable "display output" to alternate FD inside installer plugin
Instead of grepping and stripping on the caller side, we are now passing an additional file descriptor to the plugin, and it writes the "short" display output (just package names and versions, with "- " prefixed) of package installs and extension activations for the buildpack to print unchanged.

The regular full output from Composer and the plugin still goes to install.log.

GUS-W-18472789