Offer snapshot creation in destination if not exists in case sync.sh has nothing to do otherwise
Useful e.g. if obsolete formulae (that were never built in the destination) got removed, say, on stack EOL. Or if the way the snapshot hash is computed ever changes, but not the formulae themselves. Both of these of course require manual copying of the packages-SNAPSHOT.json files in the source bucket, but their existence is already guaranteed by the CI rules that force snapshot usage on Pull Requests.