Update polyfill-nodowngrades test to actually make sense in a post-heroku-20 world
The old test would, even if our implementation broke, never install a native ext-xmlrpc anymore, because we no longer have PHP 7.

Updated to use PHP 8 and an ext-newrelic polyfill for a version that is only available until PHP 8.3. The whole thing should yield PHP 8.4, and not downgrade to 8.3 on the native extension install attempt.