Stronger isolation for bootstrapped Composer and PHP
With this change, the bootstrapped PHP version is used until the very end even for the APM auto-installs. Avoids the need for trickery around PHP_INI_SCAN_DIR to prevent already-installed APM extension startups etc, and guarantees we're running a version of PHP the installer plugin is compatible with.

Also ensures error messages for users with a custom COMPOSER value (to have a different file name than composer.json) are cleaned up correctly if platform installs fail (previously, the value of the COMPOSER environment variable was re-set during that step, causing some replacements to not match).

Moves it a little closer to how things work in CNB land, too ;)

GUS-W-18988200