Close "display output" file descriptor to prevent background process hangs
If a background process spawns during the build, e.g. an APM agent such as ScoutAPM, it inherits open file descriptors. At the end of the build, the parent process (the buildpack, or whatever the process leader is for the child process) will wait for that file descriptor to be closed, even if it is closing it on the parent side. The child process, however, may never close it, since it didn't know about it upon process start, leading to the build hanging forever.

GUS-W-18850207