Tests for conflict case in APM auto-installs
If there is no ext-blackfire or ext-newrelic for the installed platform packages (which we simulate by having an explicit conflict, but it may as well be because the extension isn't available for e.g. a newly released PHP version), a warning is emitted, which we now explicitly have a test for.