Clean up message indentation and colorization
This now allows passing an indentation level to any of the message helpers, as well as a prefix.

For the indent helper, the "do not indent first line" (used in various places) case is now a lot less confusing, since it's a flag argument instead of a magic empty-or-not-empty string.

Color is now handled correctly at line boundaries, meaning warning or error colors do not leak to the next line (where a wrapper, such as Git, might be printing its own prefix).

As a result, colors are now consistently and correctly printed on the CLI and e.g. in the build logs in Heroku Dashboard.

GUS-W-19126260