Increase Apache HTTP and Nginx keep-alive timeouts to 95 seconds (#823)
For compatibility with Heroku Router 2.0, which has an idle timeout of 90 seconds, we must set a keep-alive timeout that is slightly longer than 90 seconds. This ensures that there cannot be a race condition where the router selects an existing open connection to the dyno to re-use, and the dyno simultaneously closes it (standard practice for any kind of load balancer).

GUS-W-19165676