Check if jq can be invoked during build_report::setup
There are old buildpacks in the wild that attempt to add jq, but the binary does not work anymore on newer stacks, and users with that buildpack set have forgotten about it. In that case, such a broken jq has precedence on the path and would fail the build with only an obscure error message and no useful explanation.