Fail early on build/sync Docker image cache miss
There was a GitHub outage today that corrupted the cache - the keys are reserved, but no cache exists, so nothing gets restored, but the cache can also not be written.

Unfortunately, it is not possible to have the cache action fail on write failure, or check how much has been written via an output, but we can at least bail out a bit earlier in the cases where we explicitly need the cache to be restored in order to continue.